handle floating points

+++ example/src/main/java/com/example/Foo.java
@@ -4,7 +4,7 @@ import in.abaddon.ladon.Positive;
 
 public class  Foo {
     @Positive Integer positiveInt = 4;
-    @Positive float positiveFloat = 4;
+    @Positive Double positiveFloat = 4d;
 
 
     int anyNumber = 4;
